Add implementation of find all acquisition by owner. Add multiples files for limited conflicts. Files to be merged.

+++ sadoc-parent/sadoc-dao-parent/sadoc-dao-ws/src/main/java/fr/univartois/ili/sadoc/dao/services/IAcquisitionDAO.java
@@ -21,7 +21,7 @@ public interface IAcquisitionDAO extends IWebServiceDAO<Acquisition>{
 	 * @param owner
 	 * @return the List<Acquisition> that belongs to the owner
 	 */
-	List<Acquisition> findAcquisitionByOwner(int owner);
+	List<Acquisition> findAcquisitionByOwner(long owner);
 	
 	/**
 	 * Ask the Acquisition by acronym
